+++ lib/truffle.rb
@@ -3,6 +3,7 @@
 require_relative "truffle/version"
 require_relative "truffle/content"
 require_relative "truffle/stop_reason"
+require_relative "truffle/abort_signal"
 require_relative "truffle/pricing"
 require_relative "truffle/usage"
 require_relative "truffle/message"